backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain-core>=1.2.7",
    "langgraph>=1.0.6",
    "lightrag-hku>=1.4.9.11",
    "neo4j>=6.1.0",
    "networkx>=3.6.1",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
]
